+++ packages/builder/src/stores/builder/appUrls.ts
@@ -1,24 +1,7 @@
 import { derived } from "svelte/store"
 import { appStore, selectedScreen, workspaceAppStore } from "."
 import { featureFlags } from "../portal"
-
-function buildUrl({ paths, route }: { paths: string[]; route: string }) {
-  const { host, protocol } = window.location
-
-  const trimSlashes = (value: string) =>
-    value.replace(/^\/*/, "").replace(/^\/$/, "")
-  const parsedPaths = paths
-    .map(trimSlashes)
-    .filter(x => x)
-    .join("/")
-  let url = `${protocol}//${host}/${parsedPaths}`
-
-  route = trimSlashes(route)
-  if (route) {
-    url += `/#/${route}`
-  }
-  return url
-}
+import { buildLiveUrl, buildPreviewUrl } from "@/helpers/urls"
 
 export const selectedAppUrls = derived(
   [workspaceAppStore, selectedScreen, featureFlags, appStore],
@@ -31,15 +14,9 @@ export const selectedAppUrls = derived(
         ? selectedWorkspaceApp.url
         : ""
 
-    const previewUrl = buildUrl({
-      paths: [$appStore.appId, workspacePrefix],
-      route,
-    })
+    const previewUrl = buildPreviewUrl($appStore, workspacePrefix, route)
 
-    const liveUrl = buildUrl({
-      paths: [`/app${$appStore.url}`, workspacePrefix],
-      route: "", // Ignore the route for live urls
-    })
+    const liveUrl = buildLiveUrl($appStore, workspacePrefix)
     return { previewUrl, liveUrl }
   }
 )
